long doubles again

+++ libc/stdio/printf.c
@@ -705,7 +705,7 @@ static int vaprintf(const char* restrict format, int (*put)(int), unsigned max, 
 
                 if (hasPrecision)
                 {
-                    double test_i = i;
+                    long double test_i = i;
                     unsigned digits = get_exp(&test_i, 10) + 1;
                     if (digits < precision)
                     {
@@ -762,7 +762,7 @@ static int vaprintf(const char* restrict format, int (*put)(int), unsigned max, 
 
                 if (hasPrecision)
                 {
-                    double test_u = u;
+                    long double test_u = u;
                     unsigned digits = get_exp(&test_u, 10)  + 1;
                     if (digits < precision)
                     {
@@ -835,7 +835,7 @@ static int vaprintf(const char* restrict format, int (*put)(int), unsigned max, 
 
                 if (hasPrecision)
                 {
-                    double test_o = o;
+                    long double test_o = o;
                     unsigned digits = get_exp(&test_o, 8) + 1;
                     if (digits < precision)
                     {
@@ -933,7 +933,7 @@ static int vaprintf(const char* restrict format, int (*put)(int), unsigned max, 
 
                 if (hasPrecision)
                 {
-                    double test_x = x;
+                    long double test_x = x;
                     unsigned digits = get_exp(&test_x, 16) + 1;
                     if (digits < precision)
                     {
